+++ ninja-core/src/main/java/ninja/Results.java
@@ -16,6 +16,9 @@
 
 package ninja;
 
+import ninja.lifecycle.Start;
+import ninja.utils.DateUtil;
+
 /**
  * Convenience methods for the generation of Results.
  * 
@@ -27,7 +30,7 @@ package ninja;
  *
  */
 public class Results {
-
+    
     public static Result status(int statusCode) {
 
         Result result = new Result(statusCode);
@@ -106,8 +109,7 @@ public class Results {
     }
 
     public static Result html(int statusCode) {
-        Result result = status(statusCode);
-        result.contentType(Result.TEXT_HTML);
+        Result result = status(statusCode).html();
 
         return result;
     }
@@ -124,8 +126,7 @@ public class Results {
      */
     @Deprecated
     public static Result html(Object renderable) {
-        Result result = status(Result.SC_200_OK);
-        result.contentType(Result.TEXT_HTML);
+        Result result = status(Result.SC_200_OK).html();
         result.render(renderable);
 
         return result;
@@ -140,30 +141,26 @@ public class Results {
      */
     @Deprecated
     public static Result json(int statusCode) {
-        Result result = status(statusCode);
-        result.contentType(Result.APPLICATON_JSON);
+        Result result = status(statusCode).json();
 
         return result;
     }
 
     public static Result json() {
-        Result result = status(Result.SC_200_OK);
-        result.contentType(Result.APPLICATON_JSON);
+        Result result = status(Result.SC_200_OK).json();
 
         return result;
     }
 
     public static Result json(Object renderable) {
-        Result result = status(Result.SC_200_OK);
-        result.contentType(Result.APPLICATON_JSON);
+        Result result = status(Result.SC_200_OK).json();
         result.render(renderable);
 
         return result;
     }
 
-    public static Result xml() {
-        Result result = status(Result.SC_200_OK);
-        result.contentType(Result.APPLICATION_XML);
+    public static Result xml() {       
+        Result result = status(Result.SC_200_OK).xml();
 
         return result;
     }